api/requirements.txt
fastapi
uvicorn
python-dotenv
starlette
wasabi
aiohttp
requests
pypdf
spacy
python-docx
weaviate-client
openai
pydantic
python-multipart
numpy
weaviate
scikit-learn
pydantic_settings
pymysql
aiomysql
qiniu
chardet
DBUtils==3.0.3
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
email-validator
sentence-transformers==2.2.2
pinecone-client==3.0.0
PyPDF2==3.0.1

# Potential additions:
# jsonresponse (if not part of FastAPI)
# uuid (if using a non-standard version)
